+++ soundboard/src/main/java/controller/Client.java
@@ -1,4 +1,4 @@
-package soundboard;
+package controller;
 
 import java.io.BufferedReader;
 import java.io.IOException;
